requirements.txt
filelock==3.20.1
fsspec==2025.12.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.0
pandas==2.3.3
pillow==12.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
